Separate out export function
